pyproject.toml
[tool.poetry]
name = "blender-renderfarm"
version = "0.1.0"
description = ""
authors = ["Radek Goláň jr. <r.golan@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.19"
ffmpeg-python = "^0.1.16"
udp_filetransfer = "^0.4.0"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
pylint = "^2.1"
pytest-timeout = "^1.3"
"fake-bpy-module-2.80" = "^20190718.0"
